clean code and refacto
