+++ src/actions/project-actions.ts
@@ -1,0 +1,16 @@
+import Project from '@/db/models/project'
+import { IProject } from '@/types/interfaces'
+import { revalidatePath } from 'next/cache'
+import { connect } from '@/lib/db'
+
+const createProject = async (project: IProject): Promise<void> => {
+  'use server'
+  await connect()
+  const _project = new Project(project)
+  await _project.save()
+  revalidatePath('/')
+}
+
+export {
+  createProject
+}
