+++ src/app/page.tsx
@@ -1,21 +1,10 @@
 import CreateProjectDialog from '@/components/dialogs/create-project-dialog'
 import ProjectsList from '@/components/lists/projects-list'
-import Project from '@/db/models/project'
-import { connect } from '@/lib/db'
-import { IProject } from '@/types/interfaces'
-import { revalidatePath } from 'next/cache'
-import { Model } from 'mongoose'
+import { createProject } from '@/actions/project-actions'
+import { getProjects } from '@/db/services/project-service'
 
 export default async function Home (): Promise<React.ReactNode> {
-  await connect()
-  const projects = await (Project as Model<IProject>).find().lean().exec()
-
-  const createProject = async (project: IProject): Promise<void> => {
-    'use server'
-    const _project = new Project(project)
-    await _project.save()
-    revalidatePath('/')
-  }
+  const projects = await getProjects()
 
   return (
     <div className='grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]'>
